♻️ refactor: Module 클래스를 util 패키지 아래로 이동
