+++ src/main/java/lotto/util/ExceptionModule.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import java.util.HashSet;
 import java.util.List;
