+++ src/main/java/lotto/util/InteractModule.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.util;
 
 import java.text.DecimalFormat;
 import java.util.List;
